requirements.txt
flask==3.1.2
langgraph==0.6.8
langchain_google_genai==2.1.12
langchain_openai==0.3.35
pypdf==5.1.0
python-docx==1.1.2
python-pptx==1.0.2
beautifulsoup4==4.12.3
lxml==5.3.0
